Moved actions to a subfolder, fixed StyleCops.
