+++ Nojira.Admin/Actions/ListUserAction.cs
@@ -1,0 +1,43 @@
+﻿// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+namespace Nojira.Admin
+{
+    public class ListUserAction : AdminAction
+    {
+        public override bool Execute()
+        {
+            Nojira.Utils.Database.Connect(Nojira.Utils.Config.DatabasePath, Nojira.Utils.Config.DatabasePrevPath);
+
+            foreach (Nojira.Utils.Database.User user in Nojira.Utils.Database.GetAllUsers())
+            {
+                this.Status += $"{user.UserName}\n";
+            }
+
+            if (this.Status != null)
+            {
+                this.Status = this.Status.TrimEnd('\n');
+            }
+            else
+            {
+                this.Status = "No user found.";
+            }
+
+            Nojira.Utils.Database.Disconnect();
+
+            return true;
+        }
+    }
+}
